need the fill list

+++ headers/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: khbouych <khbouych@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/05 13:55:19 by khbouych          #+#    #+#             */
-/*   Updated: 2023/10/18 19:40:57 by khbouych         ###   ########.fr       */
+/*   Updated: 2023/10/21 17:17:27 by khbouych         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,6 +27,6 @@ int		ft_strlen(char *str);
 char	*ft_strjoin(char *s1, char *s2);
 char	*ft_strchr(char *s, int c);
 char	*get_next_line(int fd);
-int     ft_strncmp(const char *str1, const char *str2, size_t n);
+int     ft_strncmp(const char *str1, const char *str2, unsigned int n);
 
 #endif
